pyproject.toml
[project]
name = "startupidea-ai"
version = "0.1.0"
description = "AI Startup Idea Validator using CrewAI and Gemini 2.0 Flash"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai>=0.102.0",
    "langchain-google-genai",
    "llama-index",
    "llama-index-llms-gemini",
    "llama-index-embeddings-gemini",
    "python-dotenv",
    "lancedb<0.30.0", # Fix: Version 0.30.1 is missing Windows wheels
    "crewai-tools>=1.11.1",
    "langchain-community>=0.4.1",
]

[tool.uv]
managed = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
